+++ src/log4qt/helpers/datetime.cpp
@@ -47,7 +47,7 @@ QString DateTime::toString(const QString &rFormat) const
     if (format == QLatin1String("ABSOLUTE"))
         format = QLatin1String("HH:mm:ss.zzz");
     if (format == QLatin1String("DATE"))
-        format = QLatin1String("dd MM yyyy HH:mm:ss.zzzz");
+        format = QLatin1String("dd MM yyyy HH:mm:ss.zzz");
 
     return formatDateTime(format);
 }